--- TruncateATable/AppContext/OracleHelperContext.cs
@@ -1,43 +1,0 @@
-/****************************************************************************
-*Copyright (c) 2017  All Rights Reserved.
-*CLR版本： 4.0.30319.42000
-*机器名称：ZHUWANSU
-*公司名称：
-*命名空间：TruncateATable.GetSQL
-*文件名：  OracleHelp
-*版本号：  V1.0.0.0
-*唯一标识：2f6f98a8-9884-49b9-adc0-3b2c9314baf5
-*当前的用户域：ZHUWANSU
-*创建人：  朱皖苏
-*电子邮箱：zhuwansu@dbgo.com
-*创建时间：2017/11/17 11:11:15
-
-*描述：
-*
-*=====================================================================
-*修改标记
-*修改时间：2017/11/17 11:11:15
-*修改人： 朱皖苏
-*版本号： V1.0.0.0
-*描述：
-*
-*****************************************************************************/
-
-using Oracle.ManagedDataAccess.Client;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Data;
-using System.Text;
-
-namespace TruncateATable.Common
-{
-
-    /// <summary>
-    /// a partial class used to OracleHelper instance context 
-    /// </summary>
-    public class TruncateATableContext
-    {
-
-    }
-}
