+++ TruncateATable/TruncateATable.csproj
@@ -38,7 +38,6 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Web" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -48,8 +47,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\OracleHelper.cs" />
-    <Compile Include="GetSQL\OracleHelper.cs" />
-    <Compile Include="GetValue\OracleHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
       <AutoGen>True</AutoGen>
